refactor: move components to sub folders
