+++ src/table/components/__tests__/TableWrapper.spec.tsx
@@ -4,8 +4,8 @@ import { stardust } from '@nebula.js/stardust';
 
 import { TableContextProvider } from '../../context';
 import TableWrapper from '../TableWrapper';
-import TableBodyWrapper from '../TableBodyWrapper';
-import TableHeadWrapper from '../TableHeadWrapper';
+import TableBodyWrapper from '../body/TableBodyWrapper';
+import TableHeadWrapper from '../head/TableHeadWrapper';
 import * as handleKeyPress from '../../utils/handle-key-press';
 import * as handleScroll from '../../utils/handle-scroll';
 import {
