+++ src/table/components/body/withStyling.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import TableCell from '@mui/material/TableCell';
-import { CellHOCProps } from '../types';
+import { CellHOCProps } from '../../types';
 
 export default function withStyling(CellComponent: typeof TableCell) {
   const HOC = (props: CellHOCProps) => {
@@ -25,7 +25,7 @@ export default function withStyling(CellComponent: typeof TableCell) {
         align={align}
         tabIndex={tabIndex}
         className={`sn-table-cell ${selectedCellClass || ''}`}
-        sx={style}
+        style={style}
         onKeyDown={onKeyDown}
         onMouseDown={onMouseDown}
         onKeyUp={onKeyUp}
