+++ src/table/components/footer/FooterWrapper.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { StyledFooterWrapper } from '../styles';
-import { FooterWrapperProps } from '../types';
+import { StyledFooterWrapper } from './styles';
+import { FooterWrapperProps } from '../../types';
 
 export default function FooterWrapper({ children, theme, footerContainer }: FooterWrapperProps) {
   return footerContainer ? (
