+++ src/table/components/head/TableHeadWrapper.tsx
@@ -1,13 +1,13 @@
 import React, { memo, useEffect, useMemo, useRef } from 'react';
 import TableHead from '@mui/material/TableHead';
 
-import { useContextSelector, TableContext } from '../context';
-import { VisuallyHidden, StyledHeadRow, StyledSortLabel, StyledHeadCell } from '../styles';
-import { getHeaderStyle } from '../utils/styling-utils';
-import { handleHeadKeyDown } from '../utils/handle-key-press';
-import { handleMouseDownLabelToFocusHeadCell, handleClickToFocusHead } from '../utils/handle-click';
-import { TableHeadWrapperProps } from '../types';
-import CellText from './CellText';
+import { useContextSelector, TableContext } from '../../context';
+import { VisuallyHidden, StyledHeadRow, StyledSortLabel, StyledHeadCell } from './styles';
+import { getHeaderStyle } from '../../utils/styling-utils';
+import { handleHeadKeyDown } from '../../utils/handle-key-press';
+import { handleMouseDownLabelToFocusHeadCell, handleClickToFocusHead } from '../../utils/handle-click';
+import { TableHeadWrapperProps } from '../../types';
+import CellText from '../CellText';
 
 function TableHeadWrapper({
   rootElement,
@@ -60,7 +60,7 @@ function TableHeadWrapper({
 
           return (
             <StyledHeadCell
-              sx={headerStyle}
+              headerStyle={headerStyle}
               key={column.id}
               align={column.align}
               className="sn-table-head-cell sn-table-cell"
