+++ src/table/components/virtualized-table/__tests__/Wrapper.spec.tsx
@@ -5,7 +5,7 @@ import { stardust } from '@nebula.js/stardust';
 import Wrapper from '../Wrapper';
 import { ExtendedTheme, ExtendedTranslator, TableLayout } from '../../../../types';
 import TableContainer from '../TableContainer';
-import FooterWrapper from '../../FooterWrapper';
+import FooterWrapper from '../../footer/FooterWrapper';
 import { MAX_PAGE_SIZE } from '../constants';
 
 jest.mock('../TableContainer');
